Add "auto" provisioning type.
Automatically use hostpath if pod is on same host as zfs pool.

This addresses #85. When storage class type is set to auto,
automatically create a hostpath volume when the scheduler selects the
specified node to run the pod, otherwise fallback to using NFS.

Note this only works when volumeBindingMode is set to
WaitForFirstConsumer in the storage class. Otherwise, when set to
Immediate, volumes will be pre-provisioned before the scheduler selects a
node for the pod consuming the volume, and options.SelectedNode will be
unset.

Note that there could be unintended side-effects if multiple pods using
the volume claim are scheduled on different nodes, depending on which
pods gets scheduled first. If the first pod gets scheduled on the ZFS
host, it will automatically use a hostpath volume and node affinity will
be set so that other pods will be prevented from running on other nodes.
On the other hand, if the second pod gets scheduled on a node which is
not the ZFS host, it will use a NFS volume, and subsequent pods will
also use NFS, even if scheduled to run on the ZFS host.