Add ability to run the provisioner directly on the ZFS host.
The default is to run kubernetes-zfs-provisioner in a container and
create datasets via SSH on a remote host.
To do that, the docker image is created with the zfs and
update-permissions stubs that will both call commands on the remote host
using SSH.

Allows running kubernetes-zfs-provisioner directly on the ZFS host by
making the update-permissions script presence optional. The zfs stub is
already optional because it merely replaces the command of the same name
on the remote host. The provisionner now uses the update-permissions
executable if it's present in the current PATH, otherwise it falls back
to executing chmod g+w directly.

Fixes #130.